Disable unsupported payment methods in Stripe settings
