+++ client/settings/general-settings-section/payment-methods-list.js
@@ -1,16 +1,14 @@
 /* global wc_stripe_settings_params */
-import { __, sprintf } from '@wordpress/i18n';
+import { sprintf } from '@wordpress/i18n';
 import React, { useState } from 'react';
 import styled from '@emotion/styled';
 import classnames from 'classnames';
-import { Button } from '@wordpress/components';
 import { Icon as IconComponent, dragHandle } from '@wordpress/icons';
 import { Reorder } from 'framer-motion';
 import interpolateComponents from 'interpolate-components';
 import PaymentMethodsMap from '../../payment-methods-map';
 import PaymentMethodDescription from './payment-method-description';
-import CustomizePaymentMethod from './customize-payment-method';
-import PaymentMethodCheckbox from './payment-method-checkbox';
+import { PaymentMethod } from './payment-method';
 import {
 	useEnabledPaymentMethodIds,
 	useGetOrderedPaymentMethodIds,
@@ -80,40 +78,6 @@ const DraggableList = styled( Reorder.Group )`
 	}
 `;
 
-const ListElement = styled.li`
-	display: flex;
-	flex-wrap: nowrap;
-	gap: 16px;
-
-	@media ( min-width: 660px ) {
-		align-items: center;
-	}
-
-	&.has-overlay {
-		position: relative;
-
-		&:after {
-			content: '';
-			position: absolute;
-			// adds some spacing for the borders, so that they're not part of the opacity
-			top: 1px;
-			bottom: 1px;
-			// ensures that the info icon isn't part of the opacity
-			left: 55px;
-			right: 0;
-			background: white;
-			opacity: 0.5;
-			pointer-events: none;
-		}
-	}
-
-	button {
-		&.hide {
-			visibility: hidden;
-		}
-	}
-`;
-
 const DraggableListElement = styled( Reorder.Item )`
 	display: flex;
 	flex-wrap: nowrap;
@@ -160,10 +124,6 @@ const PaymentMethodWrapper = styled.div`
 
 const StyledFees = styled( PaymentMethodFeesPill )`
 	flex: 1 0 auto;
-`;
-
-const CustomizeButton = styled( Button )`
-	margin-left: auto;
 `;
 
 /**
@@ -232,20 +192,6 @@ const GeneralSettingsSection = ( {
 
 	const onReorder = ( newOrderedPaymentMethodIds ) => {
 		setOrderedPaymentMethodIds( newOrderedPaymentMethodIds );
-	};
-
-	const onSaveCustomization = ( method, customizationData = null ) => {
-		setCustomizationStatus( {
-			...customizationStatus,
-			[ method ]: false,
-		} );
-
-		if ( data ) {
-			onSaveChanges(
-				'individual_payment_method_settings',
-				customizationData
-			);
-		}
 	};
 
 	return isChangingDisplayOrder ? (
@@ -321,85 +267,15 @@ const GeneralSettingsSection = ( {
 					return null;
 				}
 
-				const {
-					Icon,
-					label,
-					allows_manual_capture: isAllowingManualCapture,
-				} = PaymentMethodsMap[ method ] || {};
-
-				// Skip if there are no mapped fields for the payment method.
-				if ( ! Icon || ! label ) {
-					return null;
-				}
-
-				// Remove APMs (legacy checkout) due deprecation by Stripe on Oct 31st, 2024.
-				const deprecated =
-					// eslint-disable-next-line camelcase
-					wc_stripe_settings_params.are_apms_deprecated &&
-					method !== PAYMENT_METHOD_CARD;
-
 				return (
-					<div key={ method }>
-						<ListElement
-							key={ method }
-							className={ classnames( {
-								'has-overlay':
-									! isAllowingManualCapture &&
-									isManualCaptureEnabled,
-								expanded: customizationStatus[ method ],
-							} ) }
-						>
-							<PaymentMethodCheckbox
-								id={ method }
-								label={ label }
-								isAllowingManualCapture={
-									isAllowingManualCapture
-								}
-								disabled={ deprecated }
-							/>
-							<PaymentMethodWrapper>
-								<PaymentMethodDescription
-									id={ method }
-									Icon={ Icon }
-									description={ getFormattedPaymentMethodDescription(
-										method,
-										data.account?.default_currency
-									) }
-									label={ label }
-									deprecated={ deprecated }
-								/>
-								<StyledFees id={ method } />
-							</PaymentMethodWrapper>
-							{ ! customizationStatus[ method ] && (
-								<CustomizeButton
-									variant="secondary"
-									onClick={ () =>
-										setCustomizationStatus( {
-											...customizationStatus,
-											[ method ]: true,
-										} )
-									}
-									disabled={ deprecated }
-								>
-									{ __(
-										'Customize',
-										'woocommerce-gateway-stripe'
-									) }
-								</CustomizeButton>
-							) }
-						</ListElement>
-						{ customizationStatus[ method ] && (
-							<CustomizePaymentMethod
-								method={ method }
-								onClose={ ( customizationData ) =>
-									onSaveCustomization(
-										method,
-										customizationData
-									)
-								}
-							/>
-						) }
-					</div>
+					<PaymentMethod
+						key={ method }
+						method={ method }
+						onSaveChanges={ onSaveChanges }
+						customizationStatus={ customizationStatus }
+						setCustomizationStatus={ setCustomizationStatus }
+						data={ data }
+					/>
 				);
 			} ) }
 		</List>
